+++ project/src/constants/rating-level.ts
@@ -1,0 +1,8 @@
+export enum RatingLevel {
+  UNKNOWN = 'Unknown',
+  BAD = 'Bad',
+  NORMAL = 'Normal',
+  GOOD = 'Good',
+  VERY_GOOD = 'Very good',
+  AWESOME = 'Awesome',
+}
