+++ project/src/pages/film/film.tsx
@@ -6,11 +6,12 @@ import UserBlock from '../../components/user-block/user-block';
 import { LoadingScreen } from '../../components/loading-screen/loading-screen';
 import { AppRoute } from '../../constants/routs';
 import { SimilarFilms } from '../../components/similar-films/similar-films';
-import { IconAdd, IconInList, IconPlayS } from '../../components/icon';
+import { IconPlayS } from '../../components/icon';
 import { useAppDispatch, useAppSelector } from '../../hooks';
-import { addToFavorite, fetchFilm, fetchPromo, fetchSimilarFilms } from '../../store/api-actions';
+import { fetchFilm, fetchSimilarFilms } from '../../store/api-actions';
 import { AuthorizationStatus } from '../../constants/auth';
 import { FilmDescription } from '../../components/film-description/film-description';
+import { FavoriteButton } from '../../components/favorite-button/favorite-button';
 
 function Film(): JSX.Element {
   const navigate = useNavigate();
@@ -22,14 +23,6 @@ function Film(): JSX.Element {
 
   const handlePlayButtonClick = () => {
     navigate(generatePath(AppRoute.Player, { id: String(film.data?.id) }));
-  };
-
-  const handleMyPlaylistButtonClick = async () => {
-    await dispatch(addToFavorite({
-      filmId: Number(film.data?.id),
-      status: film.data?.isFavorite ? 0 : 1,
-    }));
-    await dispatch(fetchPromo());
   };
 
   useEffect(() => {
@@ -81,14 +74,7 @@ function Film(): JSX.Element {
                   <IconPlayS />
                   <span>Play</span>
                 </button>
-                <button className="btn btn--list film-card__button" type="button" onClick={handleMyPlaylistButtonClick}>
-                  {
-                    film.data?.isFavorite ?
-                      <IconInList /> :
-                      <IconAdd />
-                  }
-                  <span>My list</span>
-                </button>
+                <FavoriteButton film={film.data} />
                 {
                   authorizationStatus === AuthorizationStatus.Auth &&
                     <Link to={generatePath(AppRoute.AddReview, { id: String(filmId) })} className="btn film-card__button">
