+++ project/src/pages/my-list/my-list.tsx
@@ -2,7 +2,7 @@ import { useEffect } from 'react';
 
 import Logo from '../../components/logo/logo';
 import UserBlock from '../../components/user-block/user-block';
-import LoadingScreen from '../../components/loading-screen/loading-screen';
+import { LoadingScreen } from '../../components/loading-screen/loading-screen';
 import { FilmList } from '../../components/film-list/film-list';
 import { useAppDispatch, useAppSelector } from '../../hooks';
 import { fetchFavorite } from '../../store/api-actions';
